Move proof functions from Merkle sm over to being standalone functions
